updated the master test file
